made files to refactor the main.rb and app.rb
